Explicitly change inactive person to anonymous
In Front middleware: Make sure that the `request.person` is set to
AnonymousContact, if the person contact for current user is inactive.
This might be redundant, if the refresh_on_logout signal is connected,
but it might not always be.

Also add a comment to help figuring out what happens.

Refs SHUUP-2236