Provide mechanism to validate uploaded wheels
This optionally enables Wheel validation for addon uploads.
To enable it, WHEEL_USER and WHEEL_PASSWORD have to be provided

WHEEL_USER should be set to the value of vk when running `wheel keygen`
on the machine used to build and sign wheels.
WHEEL_PASSWORD should be the value returned by:
    ``keyring get wheel WHEEL_USER``
where WHEEL_USER matches the value of the previously aquired vk.

It is also necessary to properly set up the keyring backend in:
~/.local/share/python_keyring/keyringrc.cfg:

```
[backend]
default-keyring=keyrings.alt.file.PlaintextKeyring
```

The same keyring used when building the wheels should be copied to:
~/.local/share/python_keyring/keyring_pass.cfg

Ref: MAR-19