Add `ProductMode.SUBSCRIPTION`
This is not a long term solution to add new `ProductMode` definitions.

In 2.0 we should consider making `ProductMode` an object instead of
this `Enum`.

Extract visible modes to more clear internal function:
This is something I stumbled into and decided to "fix". It also seems
that the visible product modes are pretty much all the product modes,
so one could arque the need for this kind of functionality.

Changes:
* Add new `ProductMode` called `SUBSCRIPTION`
* Add helper methods to identify subscriptions for product
* Add subscription to invalid variation modes
* Extract visible product modes to method