Display sales charts with the 12 months of data
This modifies the sales chart in the admin so that 12 months of data
of will be displayed at all times. The previous behavior did cause
rendering issues when having only a month of data.

Instead of showing data for the current year, which would contain
future data, we are displaying the past 12 months.

Refs: SH-524