Return the appropriate error code in the HTTP response
Update the default error handler to return the appropriate
HTTP status code.
The status_code variable is passed to the jinja template
for rendering the error, but the HTTP status is always
left as the default, which is 200.
This fix is applied to both the admin and front error
handlers.
The tests have been updated to no longer check for
`status_code == 200`, and now check for the appropriate
status value.
Updated the test defined error handlers to return the
appropriate error code instead of 200.