Revert static files query parameter
Revert the query parameter from static files calls, which was added in
commit b15747a1.

This is just a hack that should not be needed.  It is better to use
Django's ManifestStaticFilesStorage for generating unique file names for
the static files based on file contents.  This basically just needs a
single new entry to the settings:

    STATICFILES_STORAGE = (
        'django.contrib.staticfiles.storage.ManifestStaticFilesStorage')

See the documentation of ManifestStaticFilesStorage for details:

https://docs.djangoproject.com/en/1.8/ref/contrib/staticfiles/#manifeststaticfilesstorage