admin: Fix shop product filtering in PackageChildForm
The PackageChildForm tried to do some filtering for the shop products
based on `request.user`, but there is no `request.user` set and
therefore this resulted in a crash.

Fix the issue by passing user to the form methods in order to filter the
shops and the related shop products.

Also fix a filtering bug on superusers, since it used to limit the
queryset of the shops ONLY for superusers, but it should do the
opposite.