Admin: make text editor use Shuup media browser when adding pictures
Instead of inserting images in base64 form into the text field, use the current media browser
to handle images, enabling static files serving and caching.

This fixes cases when user adds a large image through drag & drop or even copy/paste.
When the server tried to load the page, it could break and hang forever.

The copy and paste and drag & drop was disabled as they are already implemented in media browser.

Refs POS-2514