Xtheme: fix editor menu for mobile devices
The iframe was not working properly on mobile devices.
When the menu is opened and user tries to scroll the menu, the entire page was being scrolled instead of just the iframe.
This was fixed by wrapping the iframe inside a relative positioned div.

Another solved issue was the size of the editor for mobile devices. The width of the editor was 450px but didn't
fit on small devices like iPhones. The editor will now expand to 100% and it will be cropped at 450px.

The editor toggler button was also fixed for small screens by showing it over the editor, at the very bottom,
when screen sizes are smaller than 560px (the toggle button width + editor iframe width).

Refs POS-2506