Admin: fix the product images management form part
The main issue is that AJAX is used to add the files uploaded to the
current product media. It already generates the many-to-many relations
in the backend and the formset forms don't contain the ids of the
ProductMedia instances created.

To solve the issue, we return the ids of the ProductMedia created in the
backend and attach those to the hidden control inputs. Each ProductMedia
will have a corresponding hiden control input in the frontend.

This allows to set the primary image

Refs REAL-100 REAL-111
Fixes #1955