Add tests around basket line extra data
I think also passing extra_data should work since that is
what you might expect if you check the order line attribute
for this. Now you need to find the order creator "hack"
around this. Not sure if there is reason, but here is few
extra tests to show current behavior around this.