Admin: add settings to make select2 fields to load data asynchronously
It is a common request of merchants to have categories options loaded
asynchronously or on page load. To solve this issue we are introducing a
new setting `SHUUP_ADMIN_LOAD_SELECT_OBJECTS_ASYNC` that can be adjusted
per project and adjust fields according to the business model.

The setting is enabled by default and will make categories and suppliers
fields to load objects asynchronously (through AJAX requests). When
disabled, the options will be rendered on page load.

Refs HEAL-35